Make wellknow work with new infra
